+++ .github/workflows/pre_merge.yaml
@@ -17,7 +17,7 @@ jobs:
         - name: Rust tests
           run: cargo test
         - name: Rust example
-          run: cargo run --example basic_sim 200 100
+          run: cargo run --example basic_sim -- -s 200 -n 100
         - name: Rust formatting
           run: cargo fmt --all -- --check
         - name: Clippy
