+++ .gitignore
@@ -1,6 +1,6 @@
-core/db/target
-core/macros/target
-core/rust_sim/target
+crates/verbs_db/target
+crates/verbs_macros/target
+crates/verbs_rs/target
 
 /target
 .vscode
