+++ crates/verbs_rs/examples/basic_sim/basic_erc20_contract/basic_erc20.abi
