+++ crates/verbs_rs/examples/basic_sim/state.rs
@@ -1,10 +1,10 @@
 use crate::ecr20;
 use alloy_primitives::{Address, Uint, U256};
 use fastrand::Rng;
-use rust_sim::agent::{Agent, AgentSet, AgentVec, RecordedAgent, SimState};
-use rust_sim::contract::Transaction;
-use rust_sim::network::{create_call, Env};
-use rust_sim::DB;
+use verbs_rs::agent::{Agent, AgentSet, AgentVec, RecordedAgent, SimState};
+use verbs_rs::contract::Transaction;
+use verbs_rs::env::{create_call, Env};
+use verbs_rs::DB;
 
 pub struct SimpleAgent {
     pub address: Address,
