+++ crates/verbs_rs/src/contract/deployed_contract.rs
