+++ crates/verbs_rs/src/contract/structs.rs
