+++ crates/verbs_rs/src/lib.rs
@@ -1,7 +1,7 @@
 pub mod agent;
 pub mod contract;
-pub mod network;
+pub mod env;
 pub mod sim_runner;
 pub mod utils;
 
-pub use db::{ForkDb, LocalDB, Requests, DB};
+pub use verbs_db::{ForkDb, LocalDB, Requests, DB};
