+++ crates/verbs_rs/src/utils.rs
