+++ rust/Cargo.toml
@@ -16,5 +16,5 @@ fastrand.workspace = true
 
 pyo3 = { version="0.20.0", features = ["extension-module"] }
 
-rust_sim = { path = "../core/rust_sim" }
-db = { path = "../core/db" }
+verbs-rs = { path = "../crates/verbs_rs" }
+verbs-db = { path = "../crates/verbs_db" }
