+++ rust/src/sim/fork_env.rs
@@ -1,9 +1,9 @@
 use super::base_env::BaseEnv;
 use super::snapshot;
 use crate::types::{PyAddress, PyEvent, PyExecutionResult, PyRevertError};
-use db::ForkDb;
 use pyo3::prelude::*;
 use pyo3::types::PyBytes;
+use verbs_rs::ForkDb;
 
 /// Simulation environment initialised with a fork backend
 ///
